Create README.md
AQI and WQI  helps in analyzing the change in air quality and Water Quality  (improvement or degradation). AQI  and WQI informs the public about environmental conditions. It is especially useful for people suffering from illnesses aggravated or caused by air pollution or Water pollution. AQI and WQI  let us know the PHP  of water or air so that the person will have a knowledge about it.